services/board_service/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy[asyncio]
asyncmy
python-dotenv
redis
httpx
apscheduler
passlib[bcrypt]
python-multipart
aiofiles